Rescale chart
